update send order, view queue order, cancel queue order

+++ app/Http/Controllers/QueueOrderController.php
@@ -2,50 +2,95 @@
 
 namespace App\Http\Controllers;
 
+use App\Domain\Entities\Notification;
+use App\Domain\Services\CartItemService;
+use App\Domain\Services\MenuService;
+use App\Domain\Services\NotificationService;
 use App\Domain\Services\QueueOrderService;
 use App\Traits\ApiResponse;
-use Illuminate\Http\Request;
+use JWTAuth;
+use Validator;
 
 class QueueOrderController extends Controller
 {
     use ApiResponse;
 
     private $queueOrderService;
+    private $menuService;
+    private $cartItemService;
+    private $notificationService;
 
     /**
      * QueueOrderController constructor.
-     * @param $queueOrderService
+     * @param QueueOrderService $queueOrderService
+     * @param MenuService $menuService
+     * @param CartItemService $cartItemService
+     * @param NotificationService $notificationService
      */
-    public function __construct(QueueOrderService $queueOrderService)
+    public function __construct(QueueOrderService $queueOrderService, MenuService $menuService, CartItemService $cartItemService, NotificationService $notificationService)
     {
         $this->queueOrderService = $queueOrderService;
+        $this->menuService = $menuService;
+        $this->cartItemService = $cartItemService;
+        $this->notificationService = $notificationService;
     }
+
 
     public function sendOrder()
     {
-        $param = request()->all();
-        $check = $this->queueOrderService->checkExistQueueOrderInTable($param['table_id']);
-        if (!$check) {
-            $data = $this->queueOrderService->insertToQueueOrder($param);
-            return $this->successResponse($data, 'Success');
+        $user = JWTAuth::user();
+        $re = [Notification::RECEIVER_WAITER];
+        $tableID = $user->_id;
+        $cartItem = $this->cartItemService->getCartItemByTableID($tableID)->toArray();
+        if ($cartItem != []) {
+            $check = $this->queueOrderService->checkExistQueueOrderInTable($tableID);
+            if (!$check) {
+                $data = $this->queueOrderService->insertToQueueOrder();
+                $this->cartItemService->deleteAllItemByTableID($tableID);
+                $this->notificationService->notification(null, Notification::TITLE_SEND_ORDER_VN, Notification::TITLE_SEND_ORDER_EN, $user, $re);
+                return $this->successResponse($data, 'Send order success');
+            } else {
+                return $this->errorResponse('Table exist queue order', null, false, 405);
+            }
+
+        } else {
+            return $this->errorResponse('Not found item to send', null, false, 404);
         }
 
-        return $this->errorResponse('Table exist queue order', null, false, 405);
     }
 
     public function getQueueOrderByTableID()
     {
         $param = request()->all();
-        $data = $this->queueOrderService->getQueueOrderByTableID($param['table_id']);
+        $validator = Validator::make($param, [
+            'table_id' => 'required'
+        ]);
+        if ($validator->fails()) {
+            return $this->errorResponse($validator->errors(), null, false, 404);
+        }
+        $tableID = $param['table_id'];
+        $queueOrder = $this->queueOrderService->getQueueOrderByTableID($tableID);
+        if ($queueOrder) {
+            return $this->successResponse($queueOrder, 'Success');
+        } else {
+            return $this->errorResponse('Not found queue order', null, false, 404);
+        }
 
-        return $this->successResponse($data, 'Success');
     }
 
-    public function cancelQueueOrder(){
+    public function cancelQueueOrder()
+    {
         $param = request()->all();
+        $validator = Validator::make($param, [
+            'table_id' => 'required'
+        ]);
+        if ($validator->fails()) {
+            return $this->errorResponse($validator->errors(), null, false, 404);
+        }
         $this->queueOrderService->delete($param['id']);
 
-        return $this->successResponse('', 'Success');
+        return $this->successResponse(null, 'Delete Success');
     }
+
 
 }
